feat(l4): align SDK with PSRAM boot flow
Rework StarrySkyL4 C builds around the verified Flash-loaded PSRAM runtime image. Compile startup, bootloader, app, drivers, and components as ordered objects, link with the L4 PSRAM linker script, and keep QSPI timing-sensitive code at -O0.

Sync L4 board and templates to the new startup/linker contract, remove unverified L4 I2C/PS2/RCU/RTC/WDG drivers and templates, and keep asm templates on the Flash-only linker path.

Update project generation, memory reporting, Kconfig metadata, and L4 documentation for the verified st7735_clock baseline.

+++ board/StarrySkyL4/driver/archinfo/archinfo.c
@@ -3,7 +3,10 @@
 #include <stdio.h>
 
 void hal_archinfo_info(void) {
-    printf("SYS: %x IDL: %x IDH: %x\n", REG_ARCHINFO_0_SYS, REG_ARCHINFO_0_IDL, REG_ARCHINFO_0_IDH);
+    printf("SYS: %x IDL: %x IDH: %x\n",
+           (unsigned int)REG_ARCHINFO_0_SYS,
+           (unsigned int)REG_ARCHINFO_0_IDL,
+           (unsigned int)REG_ARCHINFO_0_IDH);
 }
 
 void hal_archinfo_set_sys(uint32_t val) {
